PCB made by Ramez Ghobrial during his internship at FBK
Now a real PCB exists for interconnecting the two boards.
The PCB provides the eletrical interface and does also power management given the solar panel charger and the fuel gauge.
This PCB haven't been tested and it may have problems when both solar harvester and USB charger are active. If this situation never happen the PCB should work.